chore: move ingestion platform files to a new dir
