+++ nodejs/src/ingestion/ingestion-consumer.ts
@@ -22,12 +22,12 @@ import { PromiseScheduler } from '../utils/promise-scheduler'
 import { BatchWritingGroupStore } from '../worker/ingestion/groups/batch-writing-group-store'
 import { BatchWritingPersonsStore } from '../worker/ingestion/persons/batch-writing-person-store'
 import { FlushResult, PersonsStore } from '../worker/ingestion/persons/persons-store'
-import { selectPipelineFactory } from './pipeline-selector'
-import { Config, Context, Input } from './pipelines-registry'
 import { BatchPipeline } from './pipelines/batch-pipeline.interface'
 import { newBatchPipelineBuilder } from './pipelines/builders'
 import { createContext } from './pipelines/helpers'
 import { ok } from './pipelines/results'
+import { Config, Context, Input } from './platform/registry'
+import { selectPipelineFactory } from './platform/selector'
 import { MemoryRateLimiter } from './utils/overflow-detector'
 
 /**
